requirements.txt
torch
torchvision
numpy
tqdm
Pillow
